ptr_fix
shared_ptr + unique_ptr + file parser reworked + other simple fixes

+++ Lab/MeshViewer/MeshViewer/Camera.cpp
@@ -97,7 +97,7 @@ void Camera::setDistnceToTarget(float D)
 
 void Camera::updateTarget(glm::vec3 newTarget)
 {
-	double D = distanceFromEyeToTarget();
+	float D = distanceFromEyeToTarget();
 	target = newTarget;
 	setDistnceToTarget(D);
 }
